Updated taRoutes to use cache
