+++ routes/apiRoutes.js
@@ -1,7 +1,6 @@
 const express = require('express');
-const axios = require('axios');
 const router = express.Router();
-const { getBootstrapData } = require('../cacheData');
+const { getBootstrapData } = require('../lib/fplAPIWrapper');
 
 const getGameData = (data) => {
     const currentGameweek = data.events.find(event => event.is_current).id;
@@ -12,7 +11,10 @@ const getGameData = (data) => {
 router.get('/game-data', async (req, res) => {
 
     const bootstrapData = await getBootstrapData(req);
-    const gameData = getGameData(bootstrapData.data);
+    let gameData = [];
+    if (bootstrapData.apiLive) {
+        gameData = getGameData(bootstrapData.data);
+    }
 
     res.json({ data: gameData, source: bootstrapData.source, apiLive: bootstrapData.apiLive });
 });
